Hscript timers
Added Timers dependant on the script's update cycle (so it pauses when the script does.)
Timer saving/loading
Timers will be saved to SaveData component if it exists for components, else, they are lost.
Otherwise, they are saved depending on what type of script it is.
(no way to save general purpose script timers yet)
General maintinence on components